Refactor the alternative minimum tax files Fixes #4775
