+++ changelog_entry.yaml
@@ -1,0 +1,5 @@
+- bump: patch
+  changes:
+    fixed:
+    - Refactor the alternative minimum tax file.
+    - Inculde capital gains tax in the final alternative minimum tax calculation.
