+++ policyengine_us/parameters/gov/irs/income/amt/brackets.yaml
@@ -1,57 +1,57 @@
-rates:
-  metadata:
-    unit: /1
-    reference:
-      - title: 26 U.S. Code § 55 - Alternative minimum tax imposed (b)(1)(A)
-        href: https://www.law.cornell.edu/uscode/text/26/55#b_1_A
-  1:
-    values:
+description: The IRS provides the following Alternative Minimum Tax (AMT) brackets.
+metadata:
+  type: marginal_rate
+  threshold_unit: currency-USD
+  rate_unit: /1
+  label: Alternative Minimum Tax brackets
+  reference:
+    - title: 26 U.S. Code § 55 - Alternative minimum tax imposed (b)(1)(A)
+      href: https://www.law.cornell.edu/uscode/text/26/55#b_1_A
+    - title: 2024 IRS data release
+      href: https://www.irs.gov/pub/irs-drop/rp-23-34.pdf#page=12
+    - title: 2023 IRS data release
+      href: https://www.irs.gov/pub/irs-drop/rp-22-38.pdf#page=12
+    - title: 2022 IRS data release
+      href: https://www.irs.gov/pub/irs-drop/rp-21-45.pdf#page=12
+    - title: 2021 IRS data release
+      href: https://www.irs.gov/pub/irs-drop/rp-20-45.pdf#page=12
+    - title: 2020 IRS data release
+      href: https://www.irs.gov/pub/irs-drop/rp-19-44.pdf#page=12
+    - title: CBO Tax Parameters and Effective Marginal Tax Rates | Feb 2024
+      # 1. Tax Parameters.
+      href: https://www.cbo.gov/system/files/2024-02/53724-2024-02-Tax-Parameters.xlsx
+
+brackets:
+  - threshold:
+      2013-01-01: 0
+    rate:
       2013-01-01: 0.26
-  2:
-    values:
-      2013-01-01: 0.02
-thresholds:
-  description: AMT brackets (upper thresholds)
-  metadata:
-    unit: currency-USD
-    period: year
-    uprating: gov.irs.uprating
-    reference:
-      - title: 2024 IRS data release
-        href: https://www.irs.gov/pub/irs-drop/rp-23-34.pdf#page=12
-      - title: 2023 IRS data release
-        href: https://www.irs.gov/pub/irs-drop/rp-22-38.pdf#page=12
-      - title: 2022 IRS data release
-        href: https://www.irs.gov/pub/irs-drop/rp-21-45.pdf#page=12
-      - title: 2021 IRS data release
-        href: https://www.irs.gov/pub/irs-drop/rp-20-45.pdf#page=12
-      - title: 2020 IRS data release
-        href: https://www.irs.gov/pub/irs-drop/rp-19-44.pdf#page=12
-      - title: CBO Tax Parameters and Effective Marginal Tax Rates | Feb 2024
-        # 1. Tax Parameters.
-        href: https://www.cbo.gov/system/files/2024-02/53724-2024-02-Tax-Parameters.xlsx
-  1:
-    values:
-      2013-01-01: 179_500
-      2014-01-01: 182_500
-      2015-01-01: 185_400
-      2016-01-01: 186_300
-      2017-01-01: 187_800
-      2018-01-01: 191_100
-      2019-01-01: 194_800
-      2020-01-01: 197_900
-      2021-01-01: 199_900
-      2022-01-01: 206_100
-      2023-01-01: 220_700
-      2024-01-01: 232_600
-      # CBO forecast.
-      2025-01-01: 238_500
-      2026-01-01: 243_800
-      2027-01-01: 248_800
-      2028-01-01: 253_700
-      2029-01-01: 258_600
-      2030-01-01: 263_700
-      2031-01-01: 268_900
-      2032-01-01: 274_300
-      2033-01-01: 279_700
-      2034-01-01: 285_300
+  - threshold:
+      values:
+        2013-01-01: 179_500
+        2014-01-01: 182_500
+        2015-01-01: 185_400
+        2016-01-01: 186_300
+        2017-01-01: 187_800
+        2018-01-01: 191_100
+        2019-01-01: 194_800
+        2020-01-01: 197_900
+        2021-01-01: 199_900
+        2022-01-01: 206_100
+        2023-01-01: 220_700
+        2024-01-01: 232_600
+        # CBO forecast.
+        2025-01-01: 238_500
+        2026-01-01: 243_800
+        2027-01-01: 248_800
+        2028-01-01: 253_700
+        2029-01-01: 258_600
+        2030-01-01: 263_700
+        2031-01-01: 268_900
+        2032-01-01: 274_300
+        2033-01-01: 279_700
+        2034-01-01: 285_300
+      metadata:
+        uprating: gov.irs.uprating
+    rate:
+      2013-01-01: 0.28
